refactor: service server로 하나로 병합합니다.
